DiceGame/requirements.txt
pydantic-ai-slim[groq]
python-dotenv
